projeto_mc426/requirements.txt
Django==4.2
gunicorn
whitenoise
six
django-webpush
geopy
